requirements.txt
Django==4.2.1
djangorestframework==3.14.0
Markdown==3.4.3
django-filter==23.2
django-rest-knox==4.2.0
environs==9.5.0
